splits logic of using context and expression to transform Action
